MF stable/12 r352608
Approved by:	re (gjb)

r352608:
freebsd-update: Fix src component detection

In case BASEDIR was set to a directory that differed from the default
filesystem root, freebsd-update wrongly checked for the existence
of /usr/src/COPYRIGHT to determine if the src component was
installed. Existing code to address this wasn't effective due to the
order in which configuration options were evaluated.

PR:		224048, 238558, 239997
Reviewed by:	emaste
Differential Revision:	https://reviews.freebsd.org/D21579