MFC r360071-r360072: force -fcommon for parts of kernel/kmods that need
Note that this isn't a 1:1 MFC, as a couple more instances needed to be
patched to force -fcommon in the face of a compiler that defaulted to
-fno-common.

r360071:
Allow kernel modules to build with a compiler that defaults to -fno-common

This uses the same approach as r359691.

r360072:
More fixes to build the kernel with a compiler that defaults to -fno-common

Using the same approach as the last commit for the files used by genassym.sh.